% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{knit_print.MWController}
\alias{knit_print.MWController}
\title{knit_print method for MWController object}
\usage{
knit_print.MWController(x, ...)
}
\arguments{
\item{x}{MWController object}

\item{...}{arguments passed to function knit_print}
}
\description{
knit_print method for MWController object
}
