% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwCheckbox}
\alias{mwCheckbox}
\title{Add a checkbox to a manipulateWidget gadget}
\usage{
mwCheckbox(value = FALSE, label = NULL, ...)
}
\description{
Add a checkbox to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwDateRange}},
  \code{\link{mwDate}}, \code{\link{mwNumeric}},
  \code{\link{mwPassword}}, \code{\link{mwRadio}},
  \code{\link{mwSelect}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

