% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwCheckboxGroup}
\alias{mwCheckboxGroup}
\title{Add a group of checkboxes to a manipulateWidget gadget}
\usage{
mwCheckboxGroup(choices, value = NULL, label = NULL, ...)
}
\description{
Add a group of checkboxes to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwNumeric}}, \code{\link{mwPassword}},
  \code{\link{mwRadio}}, \code{\link{mwSelect}},
  \code{\link{mwSlider}}, \code{\link{mwText}}
}

