% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwControlsUI.R
\name{mwControlsUI}
\alias{mwControlsUI}
\title{Generate controls UI of a manipulateWidget gadget}
\usage{
mwControlsUI(controlList, .dir = c("v", "h"), .n = 1, .updateBtn = FALSE)
}
\arguments{
\item{controlList}{List of input controls.}

\item{.dir}{In which direction should the controls be positioned ? "v" for vertical
and "h" for horizontal.}

\item{.n}{Number of columns if \code{.dir == "v"} else number of lines.}

\item{.updateBtn}{Should an update button be added to the controls ? If
\code{TRUE}, then the graphic is updated only when the user clicks on the
update button.}
}
\value{
A \code{shiny.tag.list} that can be used to construct a UI for a shiny
  gadget.
}
\description{
This function can be used if you want to create a custom UI for your gadget
but desire to use the controls generated by function \code{manipulateWidget}
}

