% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwDateRange}
\alias{mwDateRange}
\title{Add a date range picker to a manipulateWidget gadget}
\usage{
mwDateRange(start = NULL, end = NULL, label = NULL, ...)
}
\description{
Add a date range picker to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDate}}, \code{\link{mwNumeric}},
  \code{\link{mwPassword}}, \code{\link{mwRadio}},
  \code{\link{mwSelect}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

