% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwPassword}
\alias{mwPassword}
\title{Add a password to a manipulateWidget gadget}
\usage{
mwPassword(value = "", label = NULL, ...)
}
\description{
Add a password to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwNumeric}}, \code{\link{mwRadio}},
  \code{\link{mwSelect}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

