% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwSelect}
\alias{mwSelect}
\title{Add a Select list input to a manipulateWidget gadget}
\usage{
mwSelect(choices, value = NULL, label = NULL, ...)
}
\description{
Add a Select list input to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwNumeric}}, \code{\link{mwPassword}},
  \code{\link{mwRadio}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

