% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwSlider}
\alias{mwSlider}
\title{Add a Slider to a manipulateWidget gadget}
\usage{
mwSlider(min, max, value, label = NULL, ...)
}
\description{
Add a Slider to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwNumeric}}, \code{\link{mwPassword}},
  \code{\link{mwRadio}}, \code{\link{mwSelect}},
  \code{\link{mwText}}
}

