% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwText}
\alias{mwText}
\title{Add a text input to a manipulateWidget gadget}
\usage{
mwText(value = "", label = NULL, ...)
}
\description{
Add a text input to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwNumeric}}, \code{\link{mwPassword}},
  \code{\link{mwRadio}}, \code{\link{mwSelect}},
  \code{\link{mwSlider}}
}

