% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staticImage.R
\name{staticImage}
\alias{staticImage}
\title{Include a static image in a combinedWidgets}
\usage{
staticImage(file)
}
\arguments{
\item{file}{path of the image to include.}
}
\value{
HTML code needed to include the image in a combinedWidgets object.
}
\description{
Include a static image in a combinedWidgets
}

