% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controller.R
\name{summary.MWController}
\alias{summary.MWController}
\title{summary method for MWController object}
\usage{
\method{summary}{MWController}(object, ...)
}
\arguments{
\item{object}{MWController object}

\item{...}{Not use}
}
\description{
summary method for MWController object
}
