% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{worldEnergyUse}
\alias{worldEnergyUse}
\title{Evolution of energy use per country}
\format{
An object of class \code{data.frame} with 9375 rows and 15 columns.
}
\usage{
worldEnergyUse
}
\description{
Data.frame containing energy consumption per country from 1960 to 2014. The
data comes from the World Bank website. It contains one line per
couple(country, year) and has the following columns:
}
\details{
\itemize{
  \item country Country name
  \item iso2c Country code in two characters
  \item year Year
  \item population Population of the country
  \item energy_used_per_capita Energy used per capita in kg of oil equivalent (EG.USE.PCAP.KG.OE)
  \item energy_imported_prop Proportion of energy used that has been imported (EG.IMP.CONS.ZS)
  \item energy_fossil_prop Fossil fuel energy consumption in proportion of total consumption (EG.USE.COMM.FO.ZS)
  \item energy_used Energy consumption in kg of oil equivalent
  \item energy_fossil Fossil fuel energy consumption in kg of oil equivalent
  \item prop_world_energy_used Share of the country in the world energy consumption
  \item prop_world_energy_fossil Share of the country in the world fossil energy consumption
  \item prop_world_population Share of the country in the world population
  \item long Longitude of the country
  \item lat Lattitude of the country
  \item region Region of the country
}
}
\references{
\url{https://data.worldbank.org/indicator}
}
\author{
François Guillem \email{guillem.francois@gmail.com}
}
\keyword{datasets}
