% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwDate}
\alias{mwDate}
\title{Add a date picker to a manipulateWidget gadget}
\usage{
mwDate(value = NULL, label = NULL, ...)
}
\description{
Add a date picker to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwNumeric}},
  \code{\link{mwPassword}}, \code{\link{mwRadio}},
  \code{\link{mwSelect}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

