% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{mwNumeric}
\alias{mwNumeric}
\title{Add a numeric input to a manipulateWidget gadget}
\usage{
mwNumeric(value, label = NULL, ...)
}
\description{
Add a numeric input to a manipulateWidget gadget
}
\seealso{
Other controls: \code{\link{mwCheckbox}},
  \code{\link{mwDateRange}}, \code{\link{mwDate}},
  \code{\link{mwPassword}}, \code{\link{mwRadio}},
  \code{\link{mwSelect}}, \code{\link{mwSlider}},
  \code{\link{mwText}}
}

